refactor(query): extract memory-limit diagnostics module
